feat: Implement stateful rubric scoring system with Boolean Logic Gates
- Added calculate_rubric_score() function with deterministic scoring logic
- Updated AI prompt to output two-part |||RUBRIC||| delimiter format
- Implemented rubric parsing with Boolean checklist extraction
- Frontend captures rubric_data and saves individual q1-q6 scores
- All unit tests passed (6/6): Red Flag Override, Q1/Q2-Q6 scoring logic
- Maintains backward compatibility with existing interviews

+++ create_missing_tables.sql
@@ -18,7 +18,7 @@ CREATE POLICY "Users can view their own scans" ON resume_scans FOR SELECT USING 
 CREATE TABLE IF NOT EXISTS interviews (
     id UUID PRIMARY KEY DEFAULT uuid_generate_v4(),
     user_id UUID REFERENCES users(id),
-    average_score INT,
+    average_score DECIMAL,
     questions JSONB,
     transcript TEXT,
     created_at TIMESTAMP WITH TIME ZONE DEFAULT NOW()
